+++ iOS/Airbnb_iOS/Airbnb_iOS/FindAccomodation/View/DataSource&Delegate/FindAccomodationTableDelegate.swift
@@ -1,0 +1,14 @@
+//
+//  FindAccomodationTableDelegate.swift
+//  Airbnb_iOS
+//
+//  Created by 김한솔 on 2022/06/02.
+//
+
+import UIKit
+
+final class FindAccomodationTableDelegate: NSObject, UITableViewDelegate {
+    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
+        return 44
+    }
+}
